Finished! onClick Function and Error Checking
After looking at alot of examples and using a little bit of chatGPT I managed to get a calculator working that I understand what is happening.

Buttons now have a value, numbers and operator were set with the button.

In app AC clears the output.

= trys the output to make sure there are no errors. If there are the output is "error" if not the code outputs the answer

if any of the numbers or operator are clicked they just get added to the equation before it is analysed.